requirements.txt
flask
pyzbar
Pillow
SpeechRecognition
gunicorn